Remove steps no longer supported
