+++ MWCameraPlugin/MWCameraPlugin/ar.lproj/Localizable.strings
@@ -11,7 +11,6 @@
 "alert.error.title" = "خطأ";
 "alert.error.message" = "تم رفض إذن الكاميرا. من فضلك ، امنح حق الوصول إلى الكاميرا في تطبيق الإعدادات.";
 "qr.label" = "ركز على رمز الاستجابة السريعة";
-"barcode.label" = "ركز على الباركود";
 
 //VideoCaptureStep
 "videoCaptureStep.startRecording" = "يسجل";
